requirements.txt
Flask
flask_api
selenium
behave
firebase_admin
python-dotenv # para cargar las variables de entorno de .env
email_validator
flask-wtf
flask-testing
blinker
requests
